requirements.txt
click==8.0.1
cycler==0.10.0
dataclasses==0.8
Flask==2.0.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mpld3==0.5.5
mysql-connector-python==8.0.26
numpy==1.19.5
Pillow==8.3.2
pkg_resources==0.0.0
protobuf==3.18.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
six==1.16.0
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
